plugins/magellon_can_classifier_plugin/pyproject.toml
[project]
name = "magellon-can-classifier-plugin"
version = "0.1.0"
description = "CAN 2D-classification plugin for Magellon — competitive-Hebbian topology + MRA on a RELION particle stack."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Behdad Khoshbin", email = "b.khoshbin@gmail.com" },
]

dependencies = [
    "magellon-sdk[rabbitmq,nats,logging,config]>=2.0,<3.0",

    "fastapi",
    "uvicorn",
    "pydantic",
    "rich",
    "pyyaml",
    "python-json-logger",
    "prometheus-client",
    "prometheus-fastapi-instrumentator",
    "python-dotenv",

    "pika==1.3.2",

    # Phase 7b: add `torch`, `scikit-image`, `scikit-learn` once the
    # algorithm is vendored. Today's scaffold runs contract tests
    # against mocked ``classify_stack`` and doesn't need them.
    "mrcfile==1.5.0",
    "numpy==1.26.4",
]

[dependency-groups]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.21",
]

[tool.uv]
package = false

[tool.uv.sources]
magellon-sdk = { path = "../../magellon-sdk", editable = true }
